+++ spec/smarter_csv/simple_spec.rb
@@ -47,7 +47,7 @@ describe 'handling files with one column' do
   context 'when simple unix file with header and UTF-8 chars' do
     let(:file) { "#{fixture_path}/simple_with_header_utf8.csv" }
 
-    it 'loads the csv file without issues' do
+    it 'raises an exception' do
       expect{ data }.to raise_exception(SmarterCSV::NoColSepDetected)
     end
   end
@@ -55,7 +55,7 @@ describe 'handling files with one column' do
   context 'when simple windows file with header and UTF-8 chars' do
     let(:file) { "#{fixture_path}/simple_with_header_utf8_windows.csv" }
 
-    it 'loads the csv file without issues' do
+    it 'raises an exception' do
       expect{ data }.to raise_exception(SmarterCSV::NoColSepDetected)
     end
   end
